// packages/kobold-db/src/schemas/supabase.zod.generated.ts
export const jsonSchema: z.ZodSchema<Json> = z.lazy(() =>
  z
    .union([
      z.string(),
      z.number(),
      z.boolean(),
      z.record(z.string(), z.union([jsonSchema, z.undefined()])),
      z.array(jsonSchema),
    ])
    .nullable(),
);

export const zAction = z.object({
  actionCost: z.string().nullable(),
  autoHeighten: z.boolean(),
  baseLevel: z.number().nullable(),
  description: z.string(),
  id: z.number(),
  name: z.string(),
  rolls: jsonSchema,
  sheetRecordId: z.number().nullable(),
  tags: jsonSchema,
  type: z.string().nullable(),
});

export const zNewAction = z.object({
  actionCost: z.string().optional().nullable(),
  autoHeighten: z.boolean().optional(),
  baseLevel: z.number().optional().nullable(),
  description: z.string().optional(),
  id: z.number().optional(),
  name: z.string(),
  rolls: jsonSchema.optional(),
  sheetRecordId: z.number().optional().nullable(),
  tags: jsonSchema.optional(),
  type: z.string().optional().nullable(),
});

export const zActionUpdate = z.object({
  actionCost: z.string().optional().nullable(),
  autoHeighten: z.boolean().optional(),
  baseLevel: z.number().optional().nullable(),
  description: z.string().optional(),
  id: z.number().optional(),
  name: z.string().optional(),
  rolls: jsonSchema.optional(),
  sheetRecordId: z.number().optional().nullable(),
  tags: jsonSchema.optional(),
  type: z.string().optional().nullable(),
});

export const zActionRelationships = z.tuple([
  z.object({
    foreignKeyName: z.literal("action_sheet_record_id_fkey"),
    columns: z.tuple([z.literal("sheet_record_id")]),
    isOneToOne: z.literal(false),
    referencedRelation: z.literal("sheet_record"),
    referencedColumns: z.tuple([z.literal("id")]),
  }),
]);

export const zChannelDefaultCharacter = z.object({
  channelId: z.string(),
  characterId: z.number(),
  userId: z.string(),
});

export const zNewChannelDefaultCharacter = z.object({
  channelId: z.string(),
  characterId: z.number(),
  userId: z.string(),
});

export const zChannelDefaultCharacterUpdate = z.object({
  channelId: z.string().optional(),
  characterId: z.number().optional(),
  userId: z.string().optional(),
});

export const zChannelDefaultCharacterRelationships = z.tuple([
  z.object({
    foreignKeyName: z.literal(
      "channel_default_character_character_id_character_id_fk",
    ),
    columns: z.tuple([z.literal("character_id")]),
    isOneToOne: z.literal(false),
    referencedRelation: z.literal("character"),
    referencedColumns: z.tuple([z.literal("id")]),
  }),
]);

export const zCharacter = z.object({
  charId: z.number(),
  createdAt: z.string(),
  gameId: z.number().nullable(),
  id: z.number(),
  importSource: z.string(),
  isActiveCharacter: z.boolean(),
  lastUpdatedAt: z.string(),
  name: z.string(),
  sheetRecordId: z.number(),
  userId: z.string(),
});

export const zNewCharacter = z.object({
  charId: z.number(),
  createdAt: z.string().optional(),
  gameId: z.number().optional().nullable(),
  id: z.number().optional(),
  importSource: z.string(),
  isActiveCharacter: z.boolean().optional(),
  lastUpdatedAt: z.string().optional(),
  name: z.string(),
  sheetRecordId: z.number(),
  userId: z.string(),
});

export const zCharacterUpdate = z.object({
  charId: z.number().optional(),
  createdAt: z.string().optional(),
  gameId: z.number().optional().nullable(),
  id: z.number().optional(),
  importSource: z.string().optional(),
  isActiveCharacter: z.boolean().optional(),
  lastUpdatedAt: z.string().optional(),
  name: z.string().optional(),
  sheetRecordId: z.number().optional(),
  userId: z.string().optional(),
});
// ...
export const zGame = z.object({
  createdAt: z.string(),
  gmUserId: z.string(),
  guildId: z.string(),
  id: z.number(),
  isActive: z.boolean(),
  lastUpdatedAt: z.string(),
  name: z.string(),
});

export const zNewGame = z.object({
  createdAt: z.string().optional(),
  gmUserId: z.string(),
  guildId: z.string(),
  id: z.number().optional(),
  isActive: z.boolean().optional(),
  lastUpdatedAt: z.string().optional(),
  name: z.string(),
});

export const zGameUpdate = z.object({
  createdAt: z.string().optional(),
  gmUserId: z.string().optional(),
  guildId: z.string().optional(),
  id: z.number().optional(),
  isActive: z.boolean().optional(),
  lastUpdatedAt: z.string().optional(),
  name: z.string().optional(),
});

export const zGuildDefaultCharacter = z.object({
  characterId: z.number(),
  guildId: z.string(),
  userId: z.string(),
});

export const zNewGuildDefaultCharacter = z.object({
  characterId: z.number(),
  guildId: z.string(),
  userId: z.string(),
});

export const zGuildDefaultCharacterUpdate = z.object({
  characterId: z.number().optional(),
  guildId: z.string().optional(),
  userId: z.string().optional(),
});

export const zGuildDefaultCharacterRelationships = z.tuple([
  z.object({
    foreignKeyName: z.literal(
      "guild_default_character_character_id_character_id_fk",
    ),
    columns: z.tuple([z.literal("character_id")]),
    isOneToOne: z.literal(false),
    referencedRelation: z.literal("character"),
    referencedColumns: z.tuple([z.literal("id")]),
  }),
]);

export const zInitiative = z.object({
  channelId: z.string(),
  createdAt: z.string(),
  currentRound: z.number(),
  currentTurnGroupId: z.number().nullable(),
  gmUserId: z.string(),
  id: z.number(),
  lastUpdatedAt: z.string(),
});

export const zNewInitiative = z.object({
  channelId: z.string(),
  createdAt: z.string().optional(),
  currentRound: z.number().optional(),
  currentTurnGroupId: z.number().optional().nullable(),
  gmUserId: z.string(),
  id: z.number().optional(),
  lastUpdatedAt: z.string().optional(),
});

export const zInitiativeUpdate = z.object({
  channelId: z.string().optional(),
  createdAt: z.string().optional(),
  currentRound: z.number().optional(),
  currentTurnGroupId: z.number().optional().nullable(),
  gmUserId: z.string().optional(),
  id: z.number().optional(),
  lastUpdatedAt: z.string().optional(),
});

export const zInitiativeActor = z.object({
  characterId: z.number().nullable(),
  createdAt: z.string(),
  gameId: z.number().nullable(),
  hideStats: z.boolean(),
  id: z.number(),
  initiativeActorGroupId: z.number(),
  initiativeId: z.number(),
  lastUpdatedAt: z.string(),
  name: z.string(),
  note: z.string(),
  referenceNpcName: z.string().nullable(),
  sheetRecordId: z.number(),
  userId: z.string(),
});

export const zNewInitiativeActor = z.object({
  characterId: z.number().optional().nullable(),
  createdAt: z.string().optional(),
  gameId: z.number().optional().nullable(),
  hideStats: z.boolean().optional(),
  id: z.number().optional(),
  initiativeActorGroupId: z.number(),
  initiativeId: z.number(),
  lastUpdatedAt: z.string().optional(),
  name: z.string(),
  note: z.string().optional(),
  referenceNpcName: z.string().optional().nullable(),
  sheetRecordId: z.number(),
  userId: z.string(),
});

export const zInitiativeActorUpdate = z.object({
  characterId: z.number().optional().nullable(),
  createdAt: z.string().optional(),
  gameId: z.number().optional().nullable(),
  hideStats: z.boolean().optional(),
  id: z.number().optional(),
  initiativeActorGroupId: z.number().optional(),
  initiativeId: z.number().optional(),
  lastUpdatedAt: z.string().optional(),
  name: z.string().optional(),
  note: z.string().optional(),
  referenceNpcName: z.string().optional().nullable(),
  sheetRecordId: z.number().optional(),
  userId: z.string().optional(),
});
// ...
export const zInitiativeActorGroup = z.object({
  createdAt: z.string(),
  id: z.number(),
  initiativeId: z.number(),
  initiativeResult: z.number(),
  lastUpdatedAt: z.string(),
  name: z.string(),
  userId: z.string(),
});

export const zNewInitiativeActorGroup = z.object({
  createdAt: z.string().optional(),
  id: z.number().optional(),
  initiativeId: z.number(),
  initiativeResult: z.number(),
  lastUpdatedAt: z.string().optional(),
  name: z.string(),
  userId: z.string(),
});

export const zInitiativeActorGroupUpdate = z.object({
  createdAt: z.string().optional(),
  id: z.number().optional(),
  initiativeId: z.number().optional(),
  initiativeResult: z.number().optional(),
  lastUpdatedAt: z.string().optional(),
  name: z.string().optional(),
  userId: z.string().optional(),
});
// ...
export const zMinion = z.object({
  characterId: z.number().nullable(),
  id: z.number(),
  name: z.string(),
  sheet: jsonSchema,
  sheetRecordId: z.number().nullable(),
});

export const zNewMinion = z.object({
  characterId: z.number().optional().nullable(),
  id: z.number().optional(),
  name: z.string(),
  sheet: jsonSchema,
  sheetRecordId: z.number().optional().nullable(),
});

export const zMinionUpdate = z.object({
  characterId: z.number().optional().nullable(),
  id: z.number().optional(),
  name: z.string().optional(),
  sheet: jsonSchema.optional(),
  sheetRecordId: z.number().optional().nullable(),
});
// ...
export const zModifier = z.object({
  description: z.string(),
  id: z.number(),
  isActive: z.boolean(),
  name: z.string(),
  rollAdjustment: z.string().nullable(),
  rollTargetTags: jsonSchema,
  severity: z.number().nullable(),
  sheetAdjustments: jsonSchema,
  sheetRecordId: z.number().nullable(),
  type: z.string(),
});

export const zNewModifier = z.object({
  description: z.string().optional(),
  id: z.number().optional(),
  isActive: z.boolean().optional(),
  name: z.string(),
  rollAdjustment: z.string().optional().nullable(),
  rollTargetTags: jsonSchema.optional(),
  severity: z.number().optional().nullable(),
  sheetAdjustments: jsonSchema.optional(),
  sheetRecordId: z.number().optional().nullable(),
  type: z.string(),
});

export const zModifierUpdate = z.object({
  description: z.string().optional(),
  id: z.number().optional(),
  isActive: z.boolean().optional(),
  name: z.string().optional(),
  rollAdjustment: z.string().optional().nullable(),
  rollTargetTags: jsonSchema.optional(),
  severity: z.number().optional().nullable(),
  sheetAdjustments: jsonSchema.optional(),
  sheetRecordId: z.number().optional().nullable(),
  type: z.string().optional(),
});
// ...
export const zRollMacro = z.object({
  id: z.number(),
  macro: z.string(),
  name: z.string(),
  sheetRecordId: z.number().nullable(),
});

export const zNewRollMacro = z.object({
  id: z.number().optional(),
  macro: z.string(),
  name: z.string(),
  sheetRecordId: z.number().optional().nullable(),
});

export const zRollMacroUpdate = z.object({
  id: z.number().optional(),
  macro: z.string().optional(),
  name: z.string().optional(),
  sheetRecordId: z.number().optional().nullable(),
});

export const zRollMacroRelationships = z.tuple([
  z.object({
    foreignKeyName: z.literal("roll_macro_sheet_record_id_fkey"),
    columns: z.tuple([z.literal("sheet_record_id")]),
    isOneToOne: z.literal(false),
    referencedRelation: z.literal("sheet_record"),
    referencedColumns: z.tuple([z.literal("id")]),
  }),
]);

export const zSheetRecord = z.object({
  conditions: jsonSchema,
  id: z.number(),
  sheet: jsonSchema,
  trackerChannelId: z.string().nullable(),
  trackerGuildId: z.string().nullable(),
  trackerMessageId: z.string().nullable(),
  trackerMode: z.string().nullable(),
});

export const zNewSheetRecord = z.object({
  conditions: jsonSchema.optional(),
  id: z.number().optional(),
  sheet: jsonSchema,
  trackerChannelId: z.string().optional().nullable(),
  trackerGuildId: z.string().optional().nullable(),
  trackerMessageId: z.string().optional().nullable(),
  trackerMode: z.string().optional().nullable(),
});

export const zSheetRecordUpdate = z.object({
  conditions: jsonSchema.optional(),
  id: z.number().optional(),
  sheet: jsonSchema.optional(),
  trackerChannelId: z.string().optional().nullable(),
  trackerGuildId: z.string().optional().nullable(),
  trackerMessageId: z.string().optional().nullable(),
  trackerMode: z.string().optional().nullable(),
});

export const zUserSettings = z.object({
  defaultCompendium: z.string(),
  initStatsNotification: z.string(),
  inlineRollsDisplay: z.string(),
  rollCompactMode: z.string(),
  userId: z.string(),
});

export const zNewUserSettings = z.object({
  defaultCompendium: z.string().optional(),
  initStatsNotification: z.string().optional(),
  inlineRollsDisplay: z.string().optional(),
  rollCompactMode: z.string().optional(),
  userId: z.string(),
});

export const zUserSettingsUpdate = z.object({
  defaultCompendium: z.string().optional(),
  initStatsNotification: z.string().optional(),
  inlineRollsDisplay: z.string().optional(),
  rollCompactMode: z.string().optional(),
  userId: z.string().optional(),
});

export const zWgAuthToken = z.object({
  accessRights: z.string(),
  accessToken: z.string(),
  charId: z.number(),
  expiresAt: z.string(),
  id: z.number(),
  tokenType: z.string(),
});

export const zNewWgAuthToken = z.object({
  accessRights: z.string(),
  accessToken: z.string(),
  charId: z.number(),
  expiresAt: z.string(),
  id: z.number().optional(),
  tokenType: z.string(),
});

export const zWgAuthTokenUpdate = z.object({
  accessRights: z.string().optional(),
  accessToken: z.string().optional(),
  charId: z.number().optional(),
  expiresAt: z.string().optional(),
  id: z.number().optional(),
  tokenType: z.string().optional(),
});